Move get_timestamp to models_utils (close #144)
